move router outlet to Review page
