--- src/DashboardUI/src/pages/application/review/ReviewLayout.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from 'react-router-dom';
-
-export function ReviewLayout() {
-  return (
-    <>
-      <Outlet />
-    </>
-  );
-}
